==> code/Extract_Triplets/14.free_text_triplets.ipynb <==
import pandas as pd
import numpy as np
import pathlib
from pathlib import Path

input_path = 'D:\projects\MGKG\data\FreeText'
cluster_files_path = [file for file in Path(input_path).glob('*_cluster.csv') if Path(file).is_file()]
map_files_path = [file for file in Path(input_path).glob('*_map.txt') if Path(file).is_file()]
print(cluster_files_path)
print(map_files_path)

for i in range(4):
    cluster_df = pd.read_csv(cluster_files_path[i], sep='\t', header=0)
    map_df = pd.read_table(map_files_path[i], sep='\t', names=['head_name', 'head_MeSH', 'tail_name', 'tail_MeSH', 'path'])
    cluster_df = cluster_df['entity_pair'].str.split('|', expand=True).join(cluster_df['cluster'])
    cluster_df.rename(columns={0: 'head_MeSH', 1: 'tail_MeSH'}, inplace=True)
    df = pd.merge(cluster_df, map_df, on=['head_MeSH', 'tail_MeSH'])
    save_path = Path(str(cluster_files_path[i]).split('_cluster')[0] + '_long.csv')
    df.to_csv(save_path, sep='\t', index=0)

# 下面开始手动注释
cd_df = pd.read_csv('D:/projects/MGKG/data/FreeText/chemical_disease_long.csv', sep='\t', header=0)
cg_df = pd.read_csv('D:/projects/MGKG/data/FreeText/chemical_gene_long.csv', sep='\t', header=0)
gd_df = pd.read_csv('D:/projects/MGKG/data/FreeText/gene_disease_long.csv', sep='\t', header=0)
gg_df = pd.read_csv('D:/projects/MGKG/data/FreeText/gene_gene_long.csv', sep='\t', header=0)

# chemical_disease_dict
cd_df[cd_df['cluster']==6].sample(frac=1).drop_duplicates(subset=['head_MeSH', 'tail_MeSH']).to_csv('D:/projects/MGKG/data/FreeText/display1.csv', sep='\t', header=0, index=0)

# chemical_gene_dict
cg_df[cg_df['cluster']==9].sample(frac=1).drop_duplicates(subset=['head_MeSH', 'tail_MeSH']).to_csv('D:/projects/MGKG/data/FreeText/display2.csv', sep='\t', header=0, index=0)

# gene_disease_dict
gd_df[gd_df['cluster']==9].sample(frac=1).drop_duplicates(subset=['head_MeSH', 'tail_MeSH']).to_csv('D:/projects/MGKG/data/FreeText/display3.csv', sep='\t', header=0, index=0)

# gene_gene_dict
gg_df[gg_df['cluster']==8].sample(frac=1).drop_duplicates(subset=['head_MeSH', 'tail_MeSH']).to_csv('D:/projects/MGKG/data/FreeText/display4.csv', sep='\t', header=0, index=0)


chemical_disease_dict = {0: ['chemical:disease:T', 'chemical:disease:Pr'], 1: ['chemical:disease:T', 'chemical:disease:Pr', 'chemical:disease:Pa'], 2: ['chemical:disease:T'],
                         3: ['chemical:disease:T', 'chemical:disease:C'], 4: ['chemical:disease:Sa'], 5: ['chemical:disease:T'], 6: ['chemical:disease:T']}
chemical_gene_dict = {0: ['chemical:gene:N'], 1: ['chemical:gene:N'], 2: ['chemical:gene:N'], 3: ['chemical:gene:E+', 'chemical:gene:B'], 4: ['chemical:gene:K'], 5: ['chemical:gene:N'],
                      6: ['chemical:gene:E'], 7: ['chemical:gene:N'], 8: ['chemical:gene:A+'], 9: ['chemical:gene:A-']}
gene_disease_dict = {0: ['gene:disease:J'], 1: ['gene:disease:J'], 2: ['gene:disease:Y', 'gene:disease:U', 'gene:disease:Te'], 3: ['gene:disease:J'], 4: ['gene:disease:J'],
                     5: ['gene:disease:Te'], 6: ['gene:disease:J'], 7: ['gene:disease:Te'], 8: ['gene:disease:L'], 9: ['gene:disease:U']}
gene_gene_dict = {0: ['gene:gene:Ad'], 1: ['gene:gene:Ad'], 2: ['gene:gene:Ad'], 3: ['gene:gene:H'], 4: ['gene:gene:Ad'], 5: ['gene:gene:Ad'], 6: ['gene:gene:Rg', 'gene:gene:E'],
                  7: ['gene:gene:Ad'], 8: ['gene:gene:E+']}

dicts = [chemical_disease_dict, chemical_gene_dict, gene_disease_dict, gene_gene_dict]
dicts

for i in range(4):
    print(f'Load path is: {cluster_files_path[i]}')
    cluster_df = pd.read_csv(cluster_files_path[i], sep='\t', header=0)
    cluster_df = cluster_df['entity_pair'].str.split('|', expand=True).join(cluster_df['cluster'])
    cluster_df.rename(columns={0: 'head', 1: 'tail'}, inplace=True)
    if i == 1:
        cluster_df['tail'] = 'ENTREZ:' + cluster_df['tail']
    if i == 2:
        cluster_df['head'] = 'ENTREZ:' + cluster_df['head']
    if i == 3:
        cluster_df['head'] = 'ENTREZ:' + cluster_df['head']
        cluster_df['tail'] = 'ENTREZ:' + cluster_df['tail']
    annot_df = pd.DataFrame.from_dict(dicts[i], orient='index').reset_index()
    annot_df = pd.DataFrame(annot_df['index']).join(annot_df.drop('index', axis=1).stack().reset_index(level=1, drop=True).rename('relation'))
    match_df = pd.merge(cluster_df, annot_df, left_on='cluster', right_on='index')
    match_df.drop(['cluster', 'index'], axis=1, inplace=True)
    match_df = match_df[['head', 'relation', 'tail']]
    save_path = Path(str(cluster_files_path[i]).split('_cluster')[0] + '_triplets.csv')
    print(f'Save path is: {save_path}')
    match_df.to_csv(save_path, sep='\t', header=0, index=0)

# 为方便管理，将最后的三元组文件移动到新目录：results
